backend/requirements.txt
django
gunicorn
djangorestframework
django-cors-headers
psycopg2-binary
pandas
dj-database-url